requirements.txt
pymongo==3.2.2
requests==2.8.1
Flask==0.10.1
Flask-Cors==2.1.2
pystmark==0.4.5
gunicorn==19.6.0
bcrypt==3.1.0
PyJWT==1.4.2
mongomock==3.6.0
